chore: ts-ify gatsby pages configs, remove pathPrefix
