+++ gatsby-node.ts
@@ -1,3 +1,4 @@
 export { createPages } from './data/createPages';
-export { onCreateNode, createSchemaCustomization } from './data/onCreateNode';
+export { onCreateNode } from './data/onCreateNode';
+export { createSchemaCustomization } from './data/onCreateNode/create-graphql-schema-customization';
 export { onCreateWebpackConfig } from './gatsby-overwrite-config';
